Refactor overlap- and V1-related functions
