+++ fractal_tasks_core/lib_regions_of_interest/__init__.py
@@ -1,3 +1,6 @@
+"""
+Subpackage for ROI-related functions.
+"""
 from .v1 import *  # noqa
 from .v1_checks import *  # noqa
 from .v1_overlaps import *  # noqa
