+++ fractal_tasks_core/lib_regions_of_interest/_overlaps_common.py
@@ -10,7 +10,7 @@
 # Institute for Biomedical Research and Pelkmans Lab from the University of
 # Zurich.
 """
-Functions to identify and remove overlaps between regions of interest.
+Functions to identify overlaps between regions, not related to table specs.
 """
 import logging
 from typing import Sequence
