+++ fractal_tasks_core/lib_regions_of_interest/v1.py
@@ -10,7 +10,7 @@
 # Institute for Biomedical Research and Pelkmans Lab from the University of
 # Zurich.
 """
-Functions to handle regions of interests (via pandas and AnnData).  FIXME
+Functions to produce/process ROI tables.
 """
 import logging
 import math
